+++ zerver/views/streams.py
@@ -74,7 +74,14 @@ from zerver.lib.topic import (
 )
 from zerver.lib.typed_endpoint import ApiParamConfig, PathOnly, typed_endpoint
 from zerver.lib.typed_endpoint_validators import check_color, check_int_in_validator
-from zerver.lib.user_groups import access_user_group_for_setting
+from zerver.lib.types import AnonymousSettingGroupDict
+from zerver.lib.user_groups import (
+    GroupSettingChangeRequest,
+    access_user_group_for_setting,
+    get_group_setting_value_for_api,
+    parse_group_setting_value,
+    validate_group_setting_value_change,
+)
 from zerver.lib.users import bulk_access_users_by_email, bulk_access_users_by_id
 from zerver.lib.utils import assert_is_not_none
 from zerver.models import NamedUserGroup, Realm, Stream, UserProfile
@@ -249,9 +256,7 @@ def update_stream_backend(
     is_web_public: Json[bool] | None = None,
     new_name: str | None = None,
     message_retention_days: Json[str] | Json[int] | None = None,
-    can_remove_subscribers_group_id: Annotated[
-        Json[int | None], ApiParamConfig("can_remove_subscribers_group")
-    ] = None,
+    can_remove_subscribers_group: Json[GroupSettingChangeRequest] | None = None,
 ) -> HttpResponse:
     # We allow realm administrators to update the stream name and
     # description even for private streams.
@@ -379,28 +384,42 @@ def update_stream_backend(
 
     for setting_name, permission_configuration in Stream.stream_permission_group_settings.items():
         request_settings_dict = locals()
-        setting_group_id_name = permission_configuration.id_field_name
-
-        if setting_group_id_name not in request_settings_dict:  # nocoverage
+        assert setting_name in request_settings_dict
+        if request_settings_dict[setting_name] is None:
             continue
 
-        if request_settings_dict[setting_group_id_name] is not None and request_settings_dict[
-            setting_group_id_name
-        ] != getattr(stream, setting_group_id_name):
+        setting_value = request_settings_dict[setting_name]
+        new_setting_value = parse_group_setting_value(setting_value.new, setting_name)
+
+        expected_current_setting_value = None
+        if setting_value.old is not None:
+            expected_current_setting_value = parse_group_setting_value(
+                setting_value.old, setting_name
+            )
+
+        current_value = getattr(stream, setting_name)
+        current_setting_api_value = get_group_setting_value_for_api(current_value)
+
+        if validate_group_setting_value_change(
+            current_setting_api_value, new_setting_value, expected_current_setting_value
+        ):
             if sub is None and stream.invite_only:
                 # Admins cannot change this setting for unsubscribed private streams.
                 raise JsonableError(_("Invalid channel ID"))
-
-            user_group_id = request_settings_dict[setting_group_id_name]
             with transaction.atomic(durable=True):
                 user_group = access_user_group_for_setting(
-                    user_group_id,
+                    new_setting_value,
                     user_profile,
                     setting_name=setting_name,
                     permission_configuration=permission_configuration,
+                    current_setting_value=current_value,
                 )
                 do_change_stream_group_based_setting(
-                    stream, setting_name, user_group, acting_user=user_profile
+                    stream,
+                    setting_name,
+                    user_group,
+                    old_setting_api_value=current_setting_api_value,
+                    acting_user=user_profile,
                 )
 
     return json_success(request)
@@ -558,9 +577,7 @@ def add_subscriptions_backend(
     ] = Stream.STREAM_POST_POLICY_EVERYONE,
     history_public_to_subscribers: Json[bool] | None = None,
     message_retention_days: Json[str] | Json[int] = RETENTION_DEFAULT,
-    can_remove_subscribers_group_id: Annotated[
-        Json[int | None], ApiParamConfig("can_remove_subscribers_group")
-    ] = None,
+    can_remove_subscribers_group: Json[int | AnonymousSettingGroupDict] | None = None,
     announce: Json[bool] = False,
     principals: Json[list[str] | list[int]] | None = None,
     authorization_errors_fatal: Json[bool] = True,
@@ -572,12 +589,15 @@ def add_subscriptions_backend(
     if principals is None:
         principals = []
 
-    if can_remove_subscribers_group_id is not None:
+    if can_remove_subscribers_group is not None:
+        setting_value = parse_group_setting_value(
+            can_remove_subscribers_group, "can_remove_subscribers_group"
+        )
         permission_configuration = Stream.stream_permission_group_settings[
             "can_remove_subscribers_group"
         ]
-        can_remove_subscribers_group = access_user_group_for_setting(
-            can_remove_subscribers_group_id,
+        can_remove_subscribers_group_value = access_user_group_for_setting(
+            setting_value,
             user_profile,
             setting_name="can_remove_subscribers_group",
             permission_configuration=permission_configuration,
@@ -586,7 +606,7 @@ def add_subscriptions_backend(
         can_remove_subscribers_group_default_name = Stream.stream_permission_group_settings[
             "can_remove_subscribers_group"
         ].default_group_name
-        can_remove_subscribers_group = NamedUserGroup.objects.get(
+        can_remove_subscribers_group_value = NamedUserGroup.objects.get(
             name=can_remove_subscribers_group_default_name,
             realm=user_profile.realm,
             is_system_group=True,
@@ -612,7 +632,7 @@ def add_subscriptions_backend(
         stream_dict_copy["message_retention_days"] = parse_message_retention_days(
             message_retention_days, Stream.MESSAGE_RETENTION_SPECIAL_VALUES_MAP
         )
-        stream_dict_copy["can_remove_subscribers_group"] = can_remove_subscribers_group
+        stream_dict_copy["can_remove_subscribers_group"] = can_remove_subscribers_group_value
 
         stream_dicts.append(stream_dict_copy)